chore: refactor gap corrector service into Papers::GapCorrector & adapt code & tests
